# src/generated/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
...
			"UserListResponse": {
				"description": "User list response with pagination",
				"properties": {
					"success": {
						"type": "boolean",
						"enum": [
							true
						],
						"nullable": false
					},
					"data": {
						"items": {
							"$ref": "#/components/schemas/UserResponse"
						},
						"type": "array"
					},
					"pagination": {
						"properties": {
							"totalPages": {
								"type": "number",
								"format": "double"
							},
							"totalCount": {
								"type": "number",
								"format": "double"
							},
							"limit": {
								"type": "number",
								"format": "double"
							},
							"page": {
								"type": "number",
								"format": "double"
							}
						},
						"required": [
							"totalPages",
							"totalCount",
							"limit",
							"page"
						],
						"type": "object"
					}
				},
				"required": [
					"success",
					"data",
					"pagination"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UserErrorResponse": {
				"description": "Error response",
				"properties": {
					"success": {
						"type": "boolean",
						"enum": [
							false
						],
						"nullable": false
					},
					"error": {
						"type": "string"
					},
					"code": {
						"type": "string"
					}
				},
				"required": [
					"success",
					"error"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"SuccessResponseWrapper_BlockedUserResponse-Array_": {
				"description": "Success response wrapper",
				"properties": {
					"success": {
						"type": "boolean",
						"enum": [
							true
						],
						"nullable": false
					},
					"data": {
						"items": {
							"$ref": "#/components/schemas/BlockedUserResponse"
						},
						"type": "array"
					},
					"message": {
						"type": "string"
					}
				},
				"required": [
					"success",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
			"SuccessResponseWrapper_UserResponse_": {
				"description": "Success response wrapper",
				"properties": {
					"success": {
						"type": "boolean",
						"enum": [
							true
						],
						"nullable": false
					},
					"data": {
						"$ref": "#/components/schemas/UserResponse"
					},
					"message": {
						"type": "string"
					}
				},
				"required": [
					"success",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {
			"bearerAuth": {
				"type": "http",
				"scheme": "bearer",
				"bearerFormat": "JWT"
			}
		}
	},
	"info": {
		"title": "goldsphere-server",
		"version": "1.0.0",
		"description": "The server including the business logic and integration for the goldsphere project",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"paths": {
		"/users": {
			"get": {
				"operationId": "GetUsers",
				"responses": {
					"200": {
						"description": "Users retrieved",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UserListResponse"
								}
							}
						}
					},
					"500": {
						"description": "Internal server error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UserErrorResponse"
								}
							}
						}
					}
				},
				"description": "List all users with pagination, filtering, and sorting",
				"summary": "Get all users",
				"tags": [
					"Users"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"description": "Page number (default: 1)",
						"in": "query",
						"name": "page",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"description": "Items per page (max: 100, default: 20)",
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"description": "Search by email",
						"in": "query",
						"name": "search",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"description": "Filter by role",
						"in": "query",
						"name": "role",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"description": "Sort field",
						"in": "query",
						"name": "sortBy",
						"required": false,
						"schema": {
							"type": "string",
							"enum": [
								"email",
								"createdAt",
								"updatedAt",
								"lastLogin"
							]
						}
					},
					{
						"description": "Sort direction",
						"in": "query",
						"name": "sortOrder",
						"required": false,
						"schema": {
							"type": "string",
							"enum": [
								"asc",
								"desc"
							]
						}
					}
				]
			},
			"post": {
				"operationId": "CreateUser",
				"responses": {
					"201": {
						"description": "User created",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/SuccessResponseWrapper_UserResponse_"
										},
										{
											"$ref": "#/components/schemas/UserErrorResponse"
										}
									]
								}
							}
						}
					},
					"400": {
						"description": "Invalid input",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UserErrorResponse"
								}
							}
						}
					},
					"409": {
						"description": "Email already exists",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UserErrorResponse"
								}
							}
						}
					}
				},
				"description": "Create a new user",
				"summary": "Create user",
				"tags": [
					"Users"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"description": "User creation data",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateUserRequest",
								"description": "User creation data"
							}
						}
					}
				}
			}
		},
		"/users/blocked": {
			"get": {
				"operationId": "GetBlockedUsers",
				"responses": {
					"200": {
						"description": "Blocked users retrieved",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/SuccessResponseWrapper_BlockedUserResponse-Array_"
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UserErrorResponse"
								}
							}
						}
					},
					"403": {
						"description": "Forbidden - Admin only",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UserErrorResponse"
								}
							}
						}
					}
				},
				"description": "Get all blocked/suspended users (Admin only)",
				"summary": "Get blocked users",
				"tags": [
					"Users"
				],
				"security": [
					{
						"bearerAuth": [
							"admin"
						]
					}
				],
				"parameters": []
			}
		},
		"/users/{id}": {
			"get": {
				"operationId": "GetUserById",
				"responses": {
					"200": {
						"description": "User retrieved",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/SuccessResponseWrapper_UserResponse_"
										},
										{
											"$ref": "#/components/schemas/UserErrorResponse"
										}
									]
								}
							}
						}
					},
					"400": {
						"description": "Invalid user ID",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UserErrorResponse"
								}
							}
						}
					},
					"404": {
						"description": "User not found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UserErrorResponse"
								}
							}
						}
					}
				},
				"description": "Get user by ID",
				"summary": "Get user by ID",
				"tags": [
					"Users"
				],
				"security": [],
				"parameters": [
					{
						"description": "User UUID",
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"put": {
				"operationId": "UpdateUser",
				"responses": {
					"200": {
						"description": "User updated",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/SuccessResponseWrapper_UserResponse_"
										},
										{
											"$ref": "#/components/schemas/UserErrorResponse"
										}
									]
								}
							}
						}
					},
					"400": {
						"description": "Invalid input",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UserErrorResponse"
								}
							}
						}
					},
					"404": {
						"description": "User not found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UserErrorResponse"
								}
							}
						}
					},
					"409": {
						"description": "Email already exists",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UserErrorResponse"
								}
							}
						}
					}
				},
				"description": "Update user",
				"summary": "Update user",
				"tags": [
					"Users"
				],
				"security": [],
				"parameters": [
					{
						"description": "User UUID",
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"description": "Update data",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateUserRequest",
								"description": "Update data"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteUser",
				"responses": {
					"200": {
						"description": "User deleted",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"properties": {
												"message": {
													"type": "string"
												},
												"success": {
													"type": "boolean",
													"enum": [
														true
													],
													"nullable": false
												}
											},
											"required": [
												"message",
												"success"
											],
											"type": "object"
										},
										{
											"$ref": "#/components/schemas/UserErrorResponse"
										}
									]
								}
							}
						}
					},
					"400": {
						"description": "Invalid user ID",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UserErrorResponse"
								}
							}
						}
					},
					"404": {
						"description": "User not found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UserErrorResponse"
								}
							}
						}
					},
					"409": {
						"description": "User has dependencies",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UserErrorResponse"
								}
							}
						}
					}
				},
				"description": "Delete user (with dependency check)",
				"summary": "Delete user",
				"tags": [
					"Users"
				],
				"security": [],
				"parameters": [
					{
						"description": "User UUID",
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/users/{id}/details": {
			"get": {
				"operationId": "GetUserDetails",
				"responses": {
					"200": {
						"description": "User details retrieved",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/UserDetailsResponse"
										},
										{
											"$ref": "#/components/schemas/UserErrorResponse"
										}
									]
								}
							}
						}
					},
					"400": {
						"description": "Invalid user ID",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UserErrorResponse"
								}
							}
						}
					},
					"404": {
						"description": "User not found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UserErrorResponse"
								}
							}
						}
					},
					"500": {
						"description": "Internal server error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UserErrorResponse"
								}
							}
						}
					}
				},
				"description": "Get user with full details (profile, address, verification)",
				"summary": "Get user details",
				"tags": [
					"Users"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"description": "User UUID",
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/users/{id}/block": {
			"post": {
				"operationId": "BlockUser",
				"responses": {
					"200": {
						"description": "User blocked",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/SuccessResponseWrapper_BlockedUserResponse_"
								}
							}
						}
					},
					"400": {
						"description": "Invalid input",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UserErrorResponse"
								}
							}
						}
					},
					"403": {
						"description": "Cannot block yourself",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UserErrorResponse"
								}
							}
						}
					},
					"404": {
						"description": "User not found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UserErrorResponse"
								}
							}
						}
					},
					"409": {
						"description": "User already blocked",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UserErrorResponse"
								}
							}
						}
					}
				},
				"description": "Block a user account (Admin only)",
				"summary": "Block user",
				"tags": [
					"Users"
				],
				"security": [
					{
						"bearerAuth": [
							"admin"
						]
					}
				],
				"parameters": [
					{
						"description": "User UUID",
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"description": "Block reason",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/BlockUserRequest",
								"description": "Block reason"
							}
						}
					}
				}
			}
		},
...
	},
	"servers": [
		{
			"url": "http://localhost:3000/api"
		}
	],
...
}